move the reference queries to the corresponding dataset
as I had only one dataset before I only had one reference_queries.txt (human ratings) in the top directory of the sparql stuff. Moved this to the corresponding dataset folder now. Had to rename it to .queries, otherwise the script would have searched for wikipedia terms in the file